requirements.txt
requests==2.9.1
python-dateutil==1.5
git+git://github.com/AlgorithmGameStats/K-Means-Clustering.git